+++ src/js/model.js
@@ -1,14 +1,12 @@
+import { API_URL } from './config';
+import { getJSON } from './helpers';
 export const state = {
   recipe: {},
 };
 
 export const loadRecipe = async function (id) {
   try {
-    const res = await fetch(
-      `https://forkify-api.herokuapp.com/api/v2/recipes/${id}`
-    );
-    const data = await res.json();
-    if (!res.ok) throw new Error(`${data.message} (${res.status})`);
+    const data = await getJSON(`${API_URL}/${id}`);
     const { recipe } = data.data;
 
     state.recipe = {
@@ -22,6 +20,7 @@ export const loadRecipe = async function (id) {
       ingredients: recipe.ingredients,
     };
   } catch (err) {
-    alert(err);
+    //Temp err handeling
+    console.error(`${err} 😪`);
   }
 };
